chore: reorganize database code structure
- Rename db/ to database/
- Rename schema.js to models.js (Drizzle ORM tables)
- Rename schemas.js to validators.js (Zod validation)
- Split queries.js into database/queries/ by domain:
  - media.js, species.js, observations.js, deployments.js
  - best-media.js, utils.js
- Update all imports to use new paths
- Update drizzle.config.js and electron-builder.yml
- Update documentation to reflect new structure
