+++ drizzle.config.js
@@ -1,8 +1,8 @@
 import { defineConfig } from 'drizzle-kit'
 
 export default defineConfig({
-  schema: './src/main/db/schema.js',
-  out: './src/main/db/migrations',
+  schema: './src/main/database/models.js',
+  out: './src/main/database/migrations',
   dialect: 'sqlite',
   verbose: true,
   strict: true
