--- src/main/queries.js
@@ -1,2647 +1,0 @@
-import {
-  getDrizzleDb,
-  deployments,
-  media,
-  observations,
-  modelRuns,
-  modelOutputs,
-  getStudyDatabase,
-  executeRawQuery
-} from './db/index.js'
-import { union } from 'drizzle-orm/sqlite-core'
-import {
-  eq,
-  and,
-  desc,
-  count,
-  sql,
-  isNotNull,
-  ne,
-  inArray,
-  gte,
-  lte,
-  isNull,
-  or,
-  notExists
-} from 'drizzle-orm'
-import log from 'electron-log'
-import { DateTime } from 'luxon'
-
-/**
- * Detect timestamp format characteristics and format a DateTime to match the original format
- * This preserves the original format (with/without milliseconds, timezone, seconds)
- * @param {DateTime} newDateTime - Luxon DateTime object with the new time
- * @param {string} originalString - Original timestamp string to match format from
- * @returns {string} - Formatted timestamp string matching original format
- */
-function formatToMatchOriginal(newDateTime, originalString) {
-  if (!originalString || !newDateTime || !newDateTime.isValid) {
-    return newDateTime?.toISO() || null
-  }
-
-  // Detect format characteristics from original string
-  const hasMilliseconds = /\.\d{3}/.test(originalString)
-  const hasTimezone = /Z|[+-]\d{2}:\d{2}$/.test(originalString)
-  const hasSeconds = /T\d{2}:\d{2}:\d{2}/.test(originalString)
-
-  // Build Luxon toISO options to match original format
-  const options = {
-    suppressMilliseconds: !hasMilliseconds,
-    suppressSeconds: !hasSeconds,
-    includeOffset: hasTimezone
-  }
-
-  let result = newDateTime.toISO(options)
-
-  // If original had no timezone indicator, remove it
-  if (!hasTimezone && result) {
-    result = result.replace(/Z|[+-]\d{2}:\d{2}$/, '')
-  }
-
-  return result
-}
-
-/**
- * Extract study ID from database path
- * Uses cross-platform path splitting (handles both / and \) to correctly
- * extract the studyId on Windows, macOS, and Linux
- * @param {string} dbPath - Path to the SQLite database
- * @returns {string} - Study ID or 'unknown' if extraction fails
- */
-export function getStudyIdFromPath(dbPath) {
-  const pathParts = dbPath.split(/[/\\]/)
-  return pathParts[pathParts.length - 2] || 'unknown'
-}
-
-/**
- * Check if a dataset uses timestamp-based linking (CamTrap DP format)
- * These datasets have NULL mediaID in all observations and link via eventStart/eventEnd ranges.
- * For these datasets, blank detection is not supported (would require slow range queries).
- * @param {object} db - Drizzle database instance
- * @returns {Promise<boolean>} - True if timestamp-based (all observations have NULL mediaID)
- */
-async function isTimestampBasedDataset(db) {
-  const result = await db
-    .select({
-      hasMediaID: sql`CASE WHEN COUNT(CASE WHEN mediaID IS NOT NULL THEN 1 END) > 0 THEN 1 ELSE 0 END`
-    })
-    .from(observations)
-    .get()
-
-  return result?.hasMediaID === 0
-}
-
-/**
- * Get species distribution from the database using Drizzle ORM
- * @param {string} dbPath - Path to the SQLite database
- * @returns {Promise<Array>} - Species distribution data
- */
-export async function getSpeciesDistribution(dbPath) {
-  const startTime = Date.now()
-  log.info(`Querying species distribution from: ${dbPath}`)
-
-  try {
-    const studyId = getStudyIdFromPath(dbPath)
-
-    const db = await getDrizzleDb(studyId, dbPath)
-
-    const result = await db
-      .select({
-        scientificName: observations.scientificName,
-        count: count(observations.observationID).as('count')
-      })
-      .from(observations)
-      .where(
-        and(
-          isNotNull(observations.scientificName),
-          ne(observations.scientificName, ''),
-          sql`(${observations.observationType} IS NULL OR ${observations.observationType} != 'blank')`
-        )
-      )
-      .groupBy(observations.scientificName)
-      .orderBy(desc(count(observations.observationID)))
-
-    const elapsedTime = Date.now() - startTime
-    log.info(`Retrieved species distribution: ${result.length} species found in ${elapsedTime}ms`)
-
-    return result
-  } catch (error) {
-    log.error(`Error querying species distribution: ${error.message}`)
-    throw error
-  }
-}
-
-/**
- * Get count of blank media (media with no observations) from the database
- * For mediaID-based datasets: counts media with no linked observations
- * For timestamp-based datasets (CamTrap DP): returns 0 (blank detection not supported)
- * @param {string} dbPath - Path to the SQLite database
- * @returns {Promise<number>} - Count of media files with no observations
- */
-export async function getBlankMediaCount(dbPath) {
-  const startTime = Date.now()
-  log.info(`Querying blank media count from: ${dbPath}`)
-
-  try {
-    const studyId = getStudyIdFromPath(dbPath)
-    const db = await getDrizzleDb(studyId, dbPath)
-
-    // Check if this is a timestamp-based dataset (CamTrap DP)
-    // These datasets use eventStart/eventEnd ranges which are slow to query
-    // Return 0 blanks for now (better than showing animals incorrectly as blanks)
-    if (await isTimestampBasedDataset(db)) {
-      const elapsedTime = Date.now() - startTime
-      log.info(`Timestamp-based dataset detected, returning 0 blanks in ${elapsedTime}ms`)
-      return 0
-    }
-
-    // For mediaID-based datasets, count media with no linked observations
-    const matchingObservations = db
-      .select({ one: sql`1` })
-      .from(observations)
-      .where(eq(observations.mediaID, media.mediaID))
-
-    const result = await db
-      .select({ count: count().as('count') })
-      .from(media)
-      .where(notExists(matchingObservations))
-      .get()
-
-    const blankCount = result?.count || 0
-    const elapsedTime = Date.now() - startTime
-    log.info(`Retrieved blank media count: ${blankCount} in ${elapsedTime}ms`)
-
-    return blankCount
-  } catch (error) {
-    log.error(`Error querying blank media count: ${error.message}`)
-    throw error
-  }
-}
-
-/**
- * Get deployment information from the database using Drizzle ORM
- * @param {string} dbPath - Path to the SQLite database
- * @returns {Promise<Array>} - Deployment data with one row per location
- */
-export async function getDeployments(dbPath) {
-  const startTime = Date.now()
-  log.info(`Querying deployments from: ${dbPath}`)
-
-  try {
-    const studyId = getStudyIdFromPath(dbPath)
-
-    const db = await getDrizzleDb(studyId, dbPath)
-
-    // Create subquery with ORDER BY to get deployments sorted by locationID and deploymentStart DESC
-    const subquery = db
-      .select({
-        locationName: deployments.locationName,
-        locationID: deployments.locationID,
-        deploymentID: deployments.deploymentID,
-        deploymentStart: deployments.deploymentStart,
-        deploymentEnd: deployments.deploymentEnd,
-        longitude: deployments.longitude,
-        latitude: deployments.latitude
-      })
-      .from(deployments)
-      .orderBy(deployments.locationID, desc(deployments.deploymentStart))
-      .as('subquery')
-
-    // Select distinct with GROUP BY to get one deployment per location
-    // SQLite returns the first row in each group based on the subquery's ORDER BY
-    const result = await db
-      .selectDistinct({
-        locationID: subquery.locationID,
-        locationName: subquery.locationName,
-        deploymentID: subquery.deploymentID,
-        deploymentStart: subquery.deploymentStart,
-        deploymentEnd: subquery.deploymentEnd,
-        longitude: subquery.longitude,
-        latitude: subquery.latitude
-      })
-      .from(subquery)
-      .groupBy(subquery.latitude, subquery.longitude)
-
-    const elapsedTime = Date.now() - startTime
-    log.info(
-      `Retrieved distinct deployments: ${result.length} unique coordinates found in ${elapsedTime}ms`
-    )
-
-    return result
-  } catch (error) {
-    log.error(`Error querying deployments: ${error.message}`)
-    throw error
-  }
-}
-
-/**
- * Get activity data (observation counts) per location over time periods
- * @param {string} dbPath - Path to the SQLite database
- * @returns {Promise<Object>} - Activity data with periods and counts per location
- */
-export async function getLocationsActivity(dbPath) {
-  const startTime = Date.now()
-  log.info(`Querying location activity from: ${dbPath}`)
-
-  try {
-    const studyId = getStudyIdFromPath(dbPath)
-
-    const db = await getDrizzleDb(studyId, dbPath)
-
-    // First get the total date range to calculate period size
-    const dateRange = await db
-      .select({
-        minDate: sql`MIN(${deployments.deploymentStart})`.as('minDate'),
-        maxDate: sql`MAX(${deployments.deploymentEnd})`.as('maxDate')
-      })
-      .from(deployments)
-      .get()
-
-    if (!dateRange) {
-      throw new Error('No deployment date range found')
-    }
-
-    // Get all locations
-    const locations = await db
-      .selectDistinct({
-        locationID: deployments.locationID,
-        locationName: deployments.locationName,
-        longitude: deployments.longitude,
-        latitude: deployments.latitude
-      })
-      .from(deployments)
-
-    // Get all observations with their location IDs and event start times
-    const observationData = await db
-      .select({
-        locationID: deployments.locationID,
-        eventID: observations.eventID,
-        eventStart: observations.eventStart
-      })
-      .from(observations)
-      .innerJoin(deployments, eq(observations.deploymentID, deployments.deploymentID))
-
-    // Process the data in JavaScript
-    const minDate = new Date(dateRange.minDate)
-    const maxDate = new Date(dateRange.maxDate)
-    const totalDays = (maxDate - minDate) / (1000 * 60 * 60 * 24)
-    const periodDays = Math.ceil(totalDays / 20)
-
-    // Generate periods
-    const periods = []
-    let currentStart = new Date(minDate)
-
-    while (currentStart < maxDate) {
-      const periodEnd = new Date(currentStart)
-      periodEnd.setDate(periodEnd.getDate() + periodDays)
-
-      periods.push({
-        start: currentStart.toISOString(),
-        end: periodEnd.toISOString()
-      })
-
-      currentStart = new Date(periodEnd)
-    }
-
-    // Create location map
-    const locationMap = new Map()
-    locations.forEach((location) => {
-      locationMap.set(location.locationID, {
-        locationID: location.locationID,
-        locationName: location.locationName,
-        longitude: location.longitude,
-        latitude: location.latitude,
-        periods: periods.map((period) => ({
-          start: period.start,
-          end: period.end,
-          count: 0
-        }))
-      })
-    })
-
-    // Count observations per location per period
-    const allCounts = []
-
-    observationData.forEach((obs) => {
-      const location = locationMap.get(obs.locationID)
-      if (!location) return
-
-      const obsDate = new Date(obs.eventStart)
-
-      for (let i = 0; i < periods.length; i++) {
-        const periodStart = new Date(periods[i].start)
-        const periodEnd = new Date(periods[i].end)
-
-        if (obsDate >= periodStart && obsDate < periodEnd) {
-          location.periods[i].count++
-          break
-        }
-      }
-    })
-
-    // Collect all non-zero counts for percentile calculation
-    locationMap.forEach((location) => {
-      location.periods.forEach((period) => {
-        if (period.count > 0) {
-          allCounts.push(period.count)
-        }
-      })
-    })
-
-    // Sort counts for percentile calculations
-    allCounts.sort((a, b) => a - b)
-
-    // Calculate 95th percentile of period counts
-    const percentile95Index = Math.floor(allCounts.length * 0.95)
-    const percentile90Count = allCounts[percentile95Index] || 1
-
-    const result = {
-      startDate: dateRange.minDate,
-      endDate: dateRange.maxDate,
-      percentile90Count,
-      locations: Array.from(locationMap.values())
-    }
-
-    const elapsedTime = Date.now() - startTime
-    log.info(
-      `Retrieved location activity data for ${result.locations.length} locations in ${elapsedTime}ms`
-    )
-    return result
-  } catch (error) {
-    log.error(`Error querying location activity: ${error.message}`)
-    throw error
-  }
-}
-
-/**
- * Get daily timeseries data for specific species
- * @param {string} dbPath - Path to the SQLite database
- * @param {Array<string>} speciesNames - List of scientific names to include
- * @returns {Promise<Object>} - Timeseries data for specified species
- */
-export async function getSpeciesTimeseries(dbPath, speciesNames = []) {
-  const startTime = Date.now()
-  log.info(`Querying species timeseries from: ${dbPath} for specific species`)
-  log.info(`Selected species: ${speciesNames.join(', ')}`)
-
-  // Check if requesting blanks (media without observations)
-  const BLANK_SENTINEL = '__blank__'
-  const requestingBlanks = speciesNames.includes(BLANK_SENTINEL)
-  const regularSpecies = speciesNames.filter((s) => s !== BLANK_SENTINEL)
-
-  try {
-    const studyId = getStudyIdFromPath(dbPath)
-    const db = await getDrizzleDb(studyId, dbPath)
-
-    // Check if this is a timestamp-based dataset (CamTrap DP)
-    // For these datasets, blank queries are not supported (would require slow range queries)
-    const isTimestampBased = requestingBlanks ? await isTimestampBasedDataset(db) : false
-    const effectiveRequestingBlanks = requestingBlanks && !isTimestampBased
-    if (requestingBlanks && isTimestampBased) {
-      log.info('Timestamp-based dataset: skipping blank timeseries')
-    }
-
-    // Prepare species filter for the complex CTE query
-    let speciesFilter = ''
-    if (regularSpecies && regularSpecies.length > 0) {
-      const quotedSpecies = regularSpecies.map((name) => `'${name.replace(/'/g, "''")}'`).join(',')
-      speciesFilter = `AND scientificName IN (${quotedSpecies})`
-    }
-
-    // Use raw SQL for complex CTE query that's difficult to express in Drizzle
-    const timeseriesQuery = `
-      WITH date_range AS (
-        SELECT
-          date(min(substr(eventStart, 1, 10)), 'weekday 0', '-7 days') AS start_week,
-          date(max(substr(eventStart, 1, 10)), 'weekday 0') AS end_week
-        FROM observations
-        WHERE substr(eventStart, 1, 4) > '1970'
-      ),
-      weeks(week_start) AS (
-        SELECT start_week FROM date_range
-        UNION ALL
-        SELECT date(week_start, '+7 days')
-        FROM weeks, date_range
-        WHERE week_start < end_week
-      ),
-      species_list AS (
-        SELECT
-          scientificName,
-          COUNT(*) as count
-        FROM observations
-        WHERE scientificName IS NOT NULL AND scientificName != ''
-        ${speciesFilter}
-        GROUP BY scientificName
-      ),
-      week_species_combinations AS (
-        SELECT
-          weeks.week_start,
-          species_list.scientificName
-        FROM weeks
-        CROSS JOIN species_list
-      ),
-      weekly_counts AS (
-        SELECT
-          date(substr(eventStart, 1, 10), 'weekday 0', '-7 days') as week_start,
-          scientificName,
-          COUNT(*) as count
-        FROM observations
-        WHERE scientificName IS NOT NULL AND scientificName != ''
-        ${speciesFilter}
-        GROUP BY week_start, scientificName
-      )
-      SELECT
-        wsc.week_start as date,
-        wsc.scientificName,
-        COALESCE(wc.count, 0) as count,
-        sl.count as total_count
-      FROM week_species_combinations wsc
-      LEFT JOIN weekly_counts wc ON wsc.week_start = wc.week_start
-        AND wsc.scientificName = wc.scientificName
-      JOIN species_list sl ON wsc.scientificName = sl.scientificName
-      ORDER BY wsc.week_start ASC, wsc.scientificName
-    `
-
-    let timeseries = []
-    let speciesData = []
-
-    // Run species query if:
-    // - Specific species are requested (regularSpecies.length > 0), OR
-    // - No species filter provided at all (speciesNames.length === 0 = get all species)
-    // Only skip when ONLY blanks are requested (regularSpecies.length === 0 AND requestingBlanks)
-    const shouldRunSpeciesQuery = regularSpecies.length > 0 || !requestingBlanks
-    if (shouldRunSpeciesQuery) {
-      timeseries = await executeRawQuery(studyId, dbPath, timeseriesQuery)
-
-      // Extract species metadata from the timeseries data
-      const speciesMap = new Map()
-      timeseries.forEach((row) => {
-        if (!speciesMap.has(row.scientificName)) {
-          speciesMap.set(row.scientificName, {
-            scientificName: row.scientificName,
-            count: row.total_count
-          })
-        }
-      })
-
-      // Convert the map to an array and sort by count descending
-      speciesData = Array.from(speciesMap.values()).sort((a, b) => b.count - a.count)
-    }
-
-    // If requesting blanks, add blank media weekly counts
-    if (effectiveRequestingBlanks) {
-      const blankTimeseriesQuery = `
-        WITH date_range AS (
-          SELECT
-            date(min(substr(timestamp, 1, 10)), 'weekday 0', '-7 days') AS start_week,
-            date(max(substr(timestamp, 1, 10)), 'weekday 0') AS end_week
-          FROM media
-          WHERE timestamp IS NOT NULL AND substr(timestamp, 1, 4) > '1970'
-        ),
-        weeks(week_start) AS (
-          SELECT start_week FROM date_range
-          UNION ALL
-          SELECT date(week_start, '+7 days')
-          FROM weeks, date_range
-          WHERE week_start < end_week
-        ),
-        blank_media AS (
-          SELECT m.mediaID, m.timestamp
-          FROM media m
-          WHERE m.timestamp IS NOT NULL
-            AND NOT EXISTS (
-              SELECT 1 FROM observations o
-              WHERE o.mediaID = m.mediaID
-            )
-        ),
-        blank_weekly_counts AS (
-          SELECT
-            date(substr(timestamp, 1, 10), 'weekday 0', '-7 days') as week_start,
-            COUNT(*) as count
-          FROM blank_media
-          GROUP BY week_start
-        ),
-        blank_total AS (
-          SELECT COUNT(*) as total_count FROM blank_media
-        )
-        SELECT
-          w.week_start as date,
-          '${BLANK_SENTINEL}' as scientificName,
-          COALESCE(bwc.count, 0) as count,
-          (SELECT total_count FROM blank_total) as total_count
-        FROM weeks w
-        LEFT JOIN blank_weekly_counts bwc ON w.week_start = bwc.week_start
-        ORDER BY w.week_start ASC
-      `
-
-      const blankTimeseries = await executeRawQuery(studyId, dbPath, blankTimeseriesQuery)
-
-      // Add blank data to timeseries
-      if (blankTimeseries.length > 0) {
-        timeseries = [...timeseries, ...blankTimeseries]
-
-        // Add blank to species data
-        const blankTotalCount = blankTimeseries[0]?.total_count || 0
-        if (blankTotalCount > 0) {
-          speciesData.push({
-            scientificName: BLANK_SENTINEL,
-            count: blankTotalCount
-          })
-        }
-      }
-    }
-
-    // Process the SQL results into the expected format
-    const processedData = processTimeseriesDataFromSql(timeseries)
-
-    const elapsedTime = Date.now() - startTime
-    log.info(
-      `Retrieved timeseries data: ${processedData.length} weeks for ${speciesData.length} species in ${elapsedTime}ms`
-    )
-    return {
-      allSpecies: speciesData,
-      timeseries: processedData
-    }
-  } catch (error) {
-    log.error(`Error querying timeseries: ${error.message}`)
-    throw error
-  }
-}
-
-/**
- * Get species geolocation data for heatmap visualization
- * @param {string} dbPath - Path to the SQLite database
- * @param {Array<string>} species - List of scientific names to include
- * @param {string} startDate - ISO date string for range start
- * @param {string} endDate - ISO date string for range end
- * @param {number} startHour - Starting hour of day (0-24)
- * @param {number} endHour - Ending hour of day (0-24)
- * @returns {Promise<Object>} - Species geolocation data for heatmap
- */
-export async function getSpeciesHeatmapData(
-  dbPath,
-  species,
-  startDate,
-  endDate,
-  startHour = 0,
-  endHour = 24,
-  includeNullTimestamps = false
-) {
-  const startTime = Date.now()
-  log.info(`Querying species heatmap data from: ${dbPath}`)
-  log.info(`Date range: ${startDate} to ${endDate}`)
-  log.info(`Time range: ${startHour} to ${endHour} hours`)
-  log.info(`Species: ${species.join(', ')}`)
-
-  try {
-    const studyId = getStudyIdFromPath(dbPath)
-
-    const db = await getDrizzleDb(studyId, dbPath)
-
-    // Build base conditions
-    const baseConditions = [
-      inArray(observations.scientificName, species),
-      isNotNull(deployments.latitude),
-      isNotNull(deployments.longitude)
-    ]
-
-    // Add date range filter with null timestamp support
-    if (includeNullTimestamps) {
-      // Include observations with null eventStart OR within date range
-      baseConditions.push(
-        or(
-          isNull(observations.eventStart),
-          and(gte(observations.eventStart, startDate), lte(observations.eventStart, endDate))
-        )
-      )
-    } else {
-      baseConditions.push(gte(observations.eventStart, startDate))
-      baseConditions.push(lte(observations.eventStart, endDate))
-    }
-
-    // Add time-of-day condition using sql template for SQLite strftime
-    if (startHour < endHour) {
-      // Simple range (e.g., 8:00 to 17:00)
-      baseConditions.push(
-        sql`CAST(strftime('%H', ${observations.eventStart}) AS INTEGER) >= ${startHour}`
-      )
-      baseConditions.push(
-        sql`CAST(strftime('%H', ${observations.eventStart}) AS INTEGER) < ${endHour}`
-      )
-    } else if (startHour > endHour) {
-      // Wrapping range (e.g., 22:00 to 6:00)
-      baseConditions.push(
-        or(
-          sql`CAST(strftime('%H', ${observations.eventStart}) AS INTEGER) >= ${startHour}`,
-          sql`CAST(strftime('%H', ${observations.eventStart}) AS INTEGER) < ${endHour}`
-        )
-      )
-    }
-    // If startHour equals endHour, we include all hours (full day)
-
-    const rows = await db
-      .select({
-        locationName: deployments.locationName,
-        latitude: deployments.latitude,
-        longitude: deployments.longitude,
-        scientificName: observations.scientificName,
-        count: count().as('count')
-      })
-      .from(observations)
-      .innerJoin(deployments, eq(observations.deploymentID, deployments.deploymentID))
-      .where(and(...baseConditions))
-      .groupBy(deployments.latitude, deployments.longitude, observations.scientificName)
-      .orderBy(desc(count()))
-
-    // Process the data to create species-specific datasets
-    const speciesData = {}
-    species.forEach((s) => {
-      speciesData[s] = []
-    })
-
-    rows.forEach((row) => {
-      if (speciesData[row.scientificName]) {
-        speciesData[row.scientificName].push({
-          lat: parseFloat(row.latitude), // Convert to number here
-          lng: parseFloat(row.longitude), // Convert to number here
-          count: row.count,
-          locationName: row.locationName
-        })
-      }
-    })
-
-    const elapsedTime = Date.now() - startTime
-    log.info(`Retrieved heatmap data: ${rows.length} location points in ${elapsedTime}ms`)
-    return speciesData
-  } catch (error) {
-    log.error(`Error querying species heatmap data: ${error.message}`)
-    throw error
-  }
-}
-
-/**
- * Get media files from the database that have animal observations with optional filtering
- * @param {string} dbPath - Path to the SQLite database
- * @param {Object} options - Query options
- * @param {number} options.limit - Maximum number of media files to return
- * @param {number} options.offset - Number of records to skip for pagination
- * @param {Array<string>} options.species - List of species to filter by (optional)
- * @param {Object} options.dateRange - Date range to filter by (optional)
- * @param {string} options.dateRange.start - Start date (ISO string)
- * @param {string} options.dateRange.end - End date (ISO string)
- * @param {Object} options.timeRange - Time of day range to filter by (optional)
- * @param {number} options.timeRange.start - Start hour (0-23)
- * @param {number} options.timeRange.end - End hour (0-23)
- * @returns {Promise<Array>} - Media files matching the criteria
- */
-export async function getMedia(dbPath, options = {}) {
-  const {
-    limit: queryLimit = 10,
-    offset: queryOffset = 0,
-    species = [],
-    dateRange = {},
-    timeRange = {},
-    includeNullTimestamps = false
-  } = options
-
-  const startTime = Date.now()
-  log.info(`Querying media files from: ${dbPath} with filtering options`)
-  log.info(`Pagination: limit ${queryLimit}, offset ${queryOffset}`)
-
-  // Check if requesting blanks (media without observations)
-  const BLANK_SENTINEL = '__blank__'
-  const requestingBlanks = species.includes(BLANK_SENTINEL)
-  const regularSpecies = species.filter((s) => s !== BLANK_SENTINEL)
-
-  if (species.length > 0) {
-    log.info(`Species filter: ${species.join(', ')}`)
-    if (requestingBlanks) {
-      log.info(`Including blank media (no observations)`)
-    }
-  }
-
-  if (dateRange.start && dateRange.end) {
-    log.info(`Date range: ${typeof dateRange.start} to ${dateRange.end}`)
-  }
-
-  if (timeRange.start !== undefined && timeRange.end !== undefined) {
-    log.info(`Time range: ${timeRange.start}:00 to ${timeRange.end}:00`)
-  }
-
-  try {
-    const studyId = getStudyIdFromPath(dbPath)
-
-    const db = await getDrizzleDb(studyId, dbPath)
-
-    // Check if this is a timestamp-based dataset (CamTrap DP)
-    // For these datasets, blank queries are not supported (would require slow range queries)
-    let effectiveRequestingBlanks = requestingBlanks
-    if (requestingBlanks && (await isTimestampBasedDataset(db))) {
-      log.info('Timestamp-based dataset: blank queries not supported')
-      if (regularSpecies.length === 0) {
-        // Only blanks requested, return empty
-        const elapsedTime = Date.now() - startTime
-        log.info(
-          `Returning empty for blanks-only query on timestamp-based dataset in ${elapsedTime}ms`
-        )
-        return []
-      }
-      // Mixed query: just return species, skip blanks
-      effectiveRequestingBlanks = false
-    }
-
-    // Build date/time conditions for media table (used for blank queries)
-    const mediaDateTimeConditions = []
-
-    // Add date range filter if provided
-    let startDate, endDate
-    if (dateRange.start && dateRange.end) {
-      startDate = dateRange.start instanceof Date ? dateRange.start.toISOString() : dateRange.start
-      endDate = dateRange.end instanceof Date ? dateRange.end.toISOString() : dateRange.end
-
-      log.info(`Formatted date range: ${startDate} to ${endDate}`)
-
-      if (includeNullTimestamps) {
-        mediaDateTimeConditions.push(
-          or(
-            isNull(media.timestamp),
-            and(gte(media.timestamp, startDate), lte(media.timestamp, endDate))
-          )
-        )
-      } else {
-        mediaDateTimeConditions.push(gte(media.timestamp, startDate))
-        mediaDateTimeConditions.push(lte(media.timestamp, endDate))
-      }
-    }
-
-    // Add time of day filter if provided
-    if (timeRange.start !== undefined && timeRange.end !== undefined) {
-      if (timeRange.start < timeRange.end) {
-        const timeCondition = and(
-          sql`CAST(strftime('%H', ${media.timestamp}) AS INTEGER) >= ${timeRange.start}`,
-          sql`CAST(strftime('%H', ${media.timestamp}) AS INTEGER) < ${timeRange.end}`
-        )
-        mediaDateTimeConditions.push(
-          includeNullTimestamps ? or(isNull(media.timestamp), timeCondition) : timeCondition
-        )
-      } else if (timeRange.start > timeRange.end) {
-        const timeCondition = or(
-          sql`CAST(strftime('%H', ${media.timestamp}) AS INTEGER) >= ${timeRange.start}`,
-          sql`CAST(strftime('%H', ${media.timestamp}) AS INTEGER) < ${timeRange.end}`
-        )
-        mediaDateTimeConditions.push(
-          includeNullTimestamps ? or(isNull(media.timestamp), timeCondition) : timeCondition
-        )
-      }
-    }
-
-    // Select fields for blank media (no observations, so NULL for observation fields)
-    const blankSelectFields = {
-      mediaID: media.mediaID,
-      filePath: media.filePath,
-      fileName: media.fileName,
-      timestamp: media.timestamp,
-      deploymentID: media.deploymentID,
-      scientificName: sql`NULL`.as('scientificName'),
-      fileMediatype: media.fileMediatype,
-      eventID: sql`NULL`.as('eventID'),
-      favorite: media.favorite
-    }
-
-    // Select fields for species query (with observation data)
-    const speciesSelectFields = {
-      mediaID: media.mediaID,
-      filePath: media.filePath,
-      fileName: media.fileName,
-      timestamp: media.timestamp,
-      deploymentID: media.deploymentID,
-      scientificName: observations.scientificName,
-      fileMediatype: media.fileMediatype,
-      eventID: observations.eventID,
-      favorite: media.favorite
-    }
-
-    // Correlated subquery for blank detection (only used for mediaID-based datasets)
-    // For timestamp-based datasets, effectiveRequestingBlanks is already false
-    const matchingObservations = db
-      .select({ one: sql`1` })
-      .from(observations)
-      .where(eq(observations.mediaID, media.mediaID))
-
-    // Case 1: Only blanks requested
-    if (effectiveRequestingBlanks && regularSpecies.length === 0) {
-      const blankConditions = [notExists(matchingObservations), ...mediaDateTimeConditions]
-
-      const blankQuery = db
-        .selectDistinct(blankSelectFields)
-        .from(media)
-        .where(and(...blankConditions))
-        .orderBy(sql`${media.timestamp} DESC NULLS LAST`)
-        .limit(queryLimit)
-        .offset(queryOffset)
-
-      const rows = await blankQuery
-
-      const elapsedTime = Date.now() - startTime
-      log.info(
-        `Retrieved ${rows.length} blank media files (offset: ${queryOffset}) in ${elapsedTime}ms`
-      )
-      return rows
-    }
-
-    // Build conditions for species query
-    const baseConditions = [
-      isNotNull(observations.scientificName),
-      ne(observations.scientificName, ''),
-      ...mediaDateTimeConditions
-    ]
-
-    // Add species filter if provided
-    if (regularSpecies.length > 0) {
-      baseConditions.push(inArray(observations.scientificName, regularSpecies))
-    }
-
-    // Branch 1: Direct mediaID link (for ML runs, Wildlife Insights, Deepfaune imports)
-    const branch1 = db
-      .selectDistinct(speciesSelectFields)
-      .from(media)
-      .innerJoin(observations, eq(media.mediaID, observations.mediaID))
-      .where(and(...baseConditions))
-
-    // Branch 2: Timestamp link (for CamTrap DP datasets where observations have NULL mediaID)
-    const branch2Conditions = [...baseConditions, isNull(observations.mediaID)]
-    const branch2 = db
-      .selectDistinct(speciesSelectFields)
-      .from(media)
-      .innerJoin(observations, eq(media.timestamp, observations.eventStart))
-      .where(and(...branch2Conditions))
-
-    // Case 2: Mixed selection (species + blanks)
-    if (effectiveRequestingBlanks && regularSpecies.length > 0) {
-      // Use notExists with the correlated subquery for blank detection
-      const blankConditions = [notExists(matchingObservations), ...mediaDateTimeConditions]
-
-      const blankQuery = db
-        .selectDistinct(blankSelectFields)
-        .from(media)
-        .where(and(...blankConditions))
-
-      // Combine species queries and blank query with UNION
-      const rows = await union(branch1, branch2, blankQuery)
-        .orderBy(sql`timestamp DESC NULLS LAST`)
-        .limit(queryLimit)
-        .offset(queryOffset)
-
-      const elapsedTime = Date.now() - startTime
-      log.info(
-        `Retrieved ${rows.length} media files (species + blanks, offset: ${queryOffset}) in ${elapsedTime}ms`
-      )
-      return rows
-    }
-
-    // Case 3: Regular species query (no blanks)
-    const rows = await union(branch1, branch2)
-      .orderBy(sql`${media.timestamp} DESC NULLS LAST`)
-      .limit(queryLimit)
-      .offset(queryOffset)
-
-    const elapsedTime = Date.now() - startTime
-    log.info(
-      `Retrieved ${rows.length} media files matching criteria (offset: ${queryOffset}) in ${elapsedTime}ms`
-    )
-    return rows
-  } catch (error) {
-    log.error(`Error querying media with observations: ${error.message}`)
-    throw error
-  }
-}
-
-/**
- * Get hourly activity data for species
- * @param {string} dbPath - Path to the SQLite database
- * @param {Array<string>} species - List of scientific names to include
- * @param {string} startDate - ISO date string for range start
- * @param {string} endDate - ISO date string for range end
- * @returns {Promise<Object>} - Hourly activity data for specified species
- */
-export async function getSpeciesDailyActivity(dbPath, species, startDate, endDate) {
-  const startTime = Date.now()
-  log.info(`Querying species daily activity from: ${dbPath}`)
-  log.info(`Date range: ${startDate} to ${endDate}`)
-  log.info(`Species: ${species.join(', ')}`)
-
-  // Check if requesting blanks (media without observations)
-  const BLANK_SENTINEL = '__blank__'
-  const requestingBlanks = species.includes(BLANK_SENTINEL)
-  const regularSpecies = species.filter((s) => s !== BLANK_SENTINEL)
-
-  try {
-    const studyId = getStudyIdFromPath(dbPath)
-
-    const db = await getDrizzleDb(studyId, dbPath)
-
-    // Check if this is a timestamp-based dataset (CamTrap DP)
-    // For these datasets, blank queries are not supported (would require slow range queries)
-    const isTimestampBased = requestingBlanks ? await isTimestampBasedDataset(db) : false
-    const effectiveRequestingBlanks = requestingBlanks && !isTimestampBased
-    if (requestingBlanks && isTimestampBased) {
-      log.info('Timestamp-based dataset: skipping blank activity data')
-    }
-
-    let rows = []
-
-    // Query regular species if any
-    if (regularSpecies.length > 0) {
-      // Use sql template for SQLite-specific hour extraction via strftime
-      const hourColumn = sql`CAST(strftime('%H', ${observations.eventStart}) AS INTEGER)`.as('hour')
-
-      rows = await db
-        .select({
-          hour: hourColumn,
-          scientificName: observations.scientificName,
-          count: count().as('count')
-        })
-        .from(observations)
-        .where(
-          and(
-            inArray(observations.scientificName, regularSpecies),
-            gte(observations.eventStart, startDate),
-            lte(observations.eventStart, endDate)
-          )
-        )
-        .groupBy(hourColumn, observations.scientificName)
-        .orderBy(hourColumn, observations.scientificName)
-    }
-
-    // Query blank media hourly distribution if requested
-    if (effectiveRequestingBlanks) {
-      const blankHourColumn = sql`CAST(strftime('%H', ${media.timestamp}) AS INTEGER)`.as('hour')
-
-      // Correlated subquery for blank detection (only used for mediaID-based datasets)
-      // For timestamp-based datasets, effectiveRequestingBlanks is already false
-      const matchingObservations = db
-        .select({ one: sql`1` })
-        .from(observations)
-        .where(eq(observations.mediaID, media.mediaID))
-
-      const blankRows = await db
-        .select({
-          hour: blankHourColumn,
-          count: count().as('count')
-        })
-        .from(media)
-        .where(
-          and(
-            notExists(matchingObservations),
-            gte(media.timestamp, startDate),
-            lte(media.timestamp, endDate)
-          )
-        )
-        .groupBy(blankHourColumn)
-        .orderBy(blankHourColumn)
-
-      // Add blank rows with the sentinel as scientificName
-      blankRows.forEach((row) => {
-        rows.push({
-          hour: row.hour,
-          scientificName: BLANK_SENTINEL,
-          count: row.count
-        })
-      })
-    }
-
-    // Process the data to create species-specific hourly patterns
-    const hourlyData = Array(24)
-      .fill()
-      .map((_, i) => ({
-        hour: i,
-        // Initialize with 0 for each species
-        ...Object.fromEntries(species.map((s) => [s, 0]))
-      }))
-
-    // Fill in the actual data from the query results
-    rows.forEach((row) => {
-      if (row.hour !== null && hourlyData[row.hour]) {
-        hourlyData[row.hour][row.scientificName] = row.count
-      }
-    })
-
-    const elapsedTime = Date.now() - startTime
-    log.info(
-      `Retrieved daily activity data: ${rows.length} hour/species combinations in ${elapsedTime}ms`
-    )
-    return hourlyData
-  } catch (error) {
-    log.error(`Error querying species daily activity data: ${error.message}`)
-    throw error
-  }
-}
-
-// Helper function to process timeseries data from SQL query
-function processTimeseriesDataFromSql(rawData) {
-  const resultMap = new Map()
-
-  // Group by date and collect counts for each species
-  rawData.forEach((entry) => {
-    if (!resultMap.has(entry.date)) {
-      resultMap.set(entry.date, {})
-    }
-    const dateEntry = resultMap.get(entry.date)
-    dateEntry[entry.scientificName] = entry.count
-  })
-
-  // Convert map to array format
-  return Array.from(resultMap.entries())
-    .map(([date, speciesCounts]) => ({
-      date,
-      ...speciesCounts
-    }))
-    .sort((a, b) => a.date.localeCompare(b.date))
-}
-
-/**
- * Create and initialize a new database for an image directory
- * @param {string} dbPath - Path for the new SQLite database
- * @returns {Promise<Object>} - Database manager instance
- */
-export async function createImageDirectoryDatabase(dbPath) {
-  log.info(`Creating new database at: ${dbPath}`)
-
-  try {
-    const studyId = getStudyIdFromPath(dbPath)
-
-    // Use the Drizzle database manager which will create the database and run migrations
-    const manager = await getStudyDatabase(studyId, dbPath)
-
-    log.info(`Successfully created database for study ${studyId} at: ${dbPath}`)
-    return manager
-  } catch (error) {
-    log.error(`Error creating database: ${error.message}`)
-    throw error
-  }
-}
-
-/**
- * Insert deployment data into the database
- * @param {Object} manager - Database manager instance
- * @param {Array} deploymentsData - Array of deployment objects
- * @returns {Promise<void>}
- */
-export async function insertDeployments(manager, deploymentsData) {
-  log.info(`Inserting ${Object.keys(deploymentsData).length} deployments into database`)
-
-  try {
-    const db = manager.getDb()
-
-    manager.transaction(() => {
-      for (const depKey of Object.keys(deploymentsData)) {
-        const dep = deploymentsData[depKey]
-        db.insert(deployments)
-          .values({
-            deploymentID: dep.deploymentID,
-            locationID: dep.locationID,
-            locationName: dep.locationName,
-            deploymentStart: dep.deploymentStart ? dep.deploymentStart.toISO() : null,
-            deploymentEnd: dep.deploymentEnd ? dep.deploymentEnd.toISO() : null,
-            latitude: dep.latitude,
-            longitude: dep.longitude
-          })
-          .run()
-      }
-    })
-
-    log.info(`Successfully inserted ${Object.keys(deploymentsData).length} deployments`)
-  } catch (error) {
-    log.error(`Error inserting deployments: ${error.message}`)
-    throw error
-  }
-}
-
-/**
- * Insert media data into the database
- * @param {Object} manager - Database manager instance
- * @param {Array} mediaData - Array of media objects
- * @returns {Promise<void>}
- */
-export async function insertMedia(manager, mediaData) {
-  log.info(`Inserting ${Object.keys(mediaData).length} media items into database`)
-
-  try {
-    const db = manager.getDb()
-
-    manager.transaction(() => {
-      let count = 0
-      for (const mediaPath of Object.keys(mediaData)) {
-        const item = mediaData[mediaPath]
-        db.insert(media)
-          .values({
-            mediaID: item.mediaID,
-            deploymentID: item.deploymentID,
-            timestamp: item.timestamp ? item.timestamp.toISO() : null,
-            filePath: item.filePath,
-            fileName: item.fileName,
-            importFolder: item.importFolder || null,
-            folderName: item.folderName || null
-          })
-          .run()
-
-        count++
-        if (count % 1000 === 0) {
-          log.info(`Inserted ${count}/${Object.keys(mediaData).length} media items`)
-        }
-      }
-    })
-
-    log.info(`Successfully inserted ${Object.keys(mediaData).length} media items`)
-  } catch (error) {
-    log.error(`Error inserting media: ${error.message}`)
-    throw error
-  }
-}
-
-/**
- * Insert observations data into the database
- * @param {Object} manager - Database manager instance
- * @param {Array} observationsData - Array of observation objects
- * @returns {Promise<void>}
- */
-export async function insertObservations(manager, observationsData) {
-  log.info(`Inserting ${observationsData.length} observations into database`)
-
-  try {
-    const db = manager.getDb()
-
-    manager.transaction(() => {
-      let count = 0
-      for (const observation of observationsData) {
-        db.insert(observations)
-          .values({
-            observationID: observation.observationID,
-            mediaID: observation.mediaID,
-            deploymentID: observation.deploymentID,
-            eventID: observation.eventID,
-            eventStart: observation.eventStart ? observation.eventStart.toISO() : null,
-            eventEnd: observation.eventEnd ? observation.eventEnd.toISO() : null,
-            scientificName: observation.scientificName,
-            commonName: observation.commonName,
-            classificationProbability:
-              observation.classificationProbability !== undefined
-                ? observation.classificationProbability
-                : null,
-            count: observation.count !== undefined ? observation.count : null
-          })
-          .run()
-
-        count++
-        if (count % 1000 === 0) {
-          log.info(`Inserted ${count}/${observationsData.length} observations`)
-        }
-      }
-    })
-
-    log.info(`Successfully inserted ${observationsData.length} observations`)
-  } catch (error) {
-    log.error(`Error inserting observations: ${error.message}`)
-    throw error
-  }
-}
-
-/**
- * Get activity data (observation counts) per deployment over time periods
- * Uses SQL-level aggregation for performance with large datasets
- * @param {string} dbPath - Path to the SQLite database
- * @returns {Promise<Object>} - Activity data with periods and counts per deployment
- */
-export async function getDeploymentsActivity(dbPath) {
-  const startTime = Date.now()
-  log.info(`Querying deployment activity from: ${dbPath}`)
-
-  try {
-    const studyId = getStudyIdFromPath(dbPath)
-
-    const db = await getDrizzleDb(studyId, dbPath)
-
-    // First get the total date range to calculate period size
-    const dateRange = await db
-      .select({
-        minDate: sql`MIN(${deployments.deploymentStart})`.as('minDate'),
-        maxDate: sql`MAX(${deployments.deploymentEnd})`.as('maxDate')
-      })
-      .from(deployments)
-      .get()
-
-    if (!dateRange || !dateRange.minDate || !dateRange.maxDate) {
-      // Return empty result if no deployments
-      return {
-        startDate: null,
-        endDate: null,
-        percentile90Count: 1,
-        deployments: []
-      }
-    }
-
-    // Calculate period boundaries
-    const minDate = new Date(dateRange.minDate)
-    const maxDate = new Date(dateRange.maxDate)
-    const totalDays = (maxDate - minDate) / (1000 * 60 * 60 * 24)
-    const periodDays = Math.max(1, Math.ceil(totalDays / 20))
-
-    // Generate periods
-    const periods = []
-    let currentStart = new Date(minDate)
-
-    while (currentStart < maxDate) {
-      const periodEnd = new Date(currentStart)
-      periodEnd.setDate(periodEnd.getDate() + periodDays)
-
-      periods.push({
-        start: currentStart.toISOString(),
-        end: periodEnd.toISOString()
-      })
-
-      currentStart = new Date(periodEnd)
-    }
-
-    // Build SQL CASE expressions for each period
-    // This aggregates observation counts at the database level
-    const periodCases = periods.map((period, i) =>
-      sql`SUM(CASE WHEN ${observations.eventStart} >= ${period.start} AND ${observations.eventStart} < ${period.end} THEN 1 ELSE 0 END)`.as(
-        `period_${i}`
-      )
-    )
-
-    // Single aggregated query: join deployments with observations and count per period
-    const aggregatedData = await db
-      .select({
-        deploymentID: deployments.deploymentID,
-        locationName: deployments.locationName,
-        locationID: deployments.locationID,
-        deploymentStart: deployments.deploymentStart,
-        deploymentEnd: deployments.deploymentEnd,
-        latitude: deployments.latitude,
-        longitude: deployments.longitude,
-        ...Object.fromEntries(periodCases.map((c, i) => [`period_${i}`, c]))
-      })
-      .from(deployments)
-      .leftJoin(observations, eq(deployments.deploymentID, observations.deploymentID))
-      .groupBy(deployments.deploymentID)
-
-    // Transform aggregated data to expected format
-    const allCounts = []
-    const deploymentsResult = aggregatedData.map((row) => {
-      const deploymentPeriods = periods.map((period, i) => {
-        const count = row[`period_${i}`] || 0
-        if (count > 0) {
-          allCounts.push(count)
-        }
-        return {
-          start: period.start,
-          end: period.end,
-          count
-        }
-      })
-
-      return {
-        deploymentID: row.deploymentID,
-        locationName: row.locationName,
-        locationID: row.locationID,
-        deploymentStart: row.deploymentStart,
-        deploymentEnd: row.deploymentEnd,
-        latitude: row.latitude,
-        longitude: row.longitude,
-        periods: deploymentPeriods
-      }
-    })
-
-    // Sort counts for percentile calculation
-    allCounts.sort((a, b) => a - b)
-
-    // Calculate 95th percentile of period counts
-    const percentile95Index = Math.floor(allCounts.length * 0.95)
-    const percentile90Count = allCounts[percentile95Index] || 1
-
-    const result = {
-      startDate: dateRange.minDate,
-      endDate: dateRange.maxDate,
-      percentile90Count,
-      deployments: deploymentsResult
-    }
-
-    const elapsedTime = Date.now() - startTime
-    log.info(
-      `Retrieved deployment activity data for ${result.deployments.length} deployments in ${elapsedTime}ms`
-    )
-    return result
-  } catch (error) {
-    log.error(`Error querying deployment activity: ${error.message}`)
-    throw error
-  }
-}
-
-/**
- * Get files data (directories with image counts and processing progress) for local/ml_run studies
- * @param {string} dbPath - Path to the SQLite database
- * @returns {Promise<Array>} - Array of directory objects with image counts and processing progress
- */
-export async function getFilesData(dbPath) {
-  const startTime = Date.now()
-  log.info(`Querying files data from: ${dbPath}`)
-
-  try {
-    const studyId = getStudyIdFromPath(dbPath)
-
-    const db = await getDrizzleDb(studyId, dbPath)
-
-    // Query to get directory statistics with most recent model used
-    const rows = await db
-      .select({
-        folderName: media.folderName,
-        importFolder: media.importFolder,
-        imageCount: count(sql`CASE WHEN ${media.fileMediatype} NOT LIKE 'video/%' THEN 1 END`).as(
-          'imageCount'
-        ),
-        videoCount: count(sql`CASE WHEN ${media.fileMediatype} LIKE 'video/%' THEN 1 END`).as(
-          'videoCount'
-        ),
-        processedCount: count(observations.observationID).as('processedCount'),
-        lastModelUsed: sql`(
-          SELECT mr.modelID || ' ' || mr.modelVersion
-          FROM model_outputs mo
-          INNER JOIN media m2 ON mo.mediaID = m2.mediaID
-          INNER JOIN model_runs mr ON mo.runID = mr.id
-          WHERE m2.folderName = ${media.folderName}
-          ORDER BY mr.startedAt DESC
-          LIMIT 1
-        )`.as('lastModelUsed')
-      })
-      .from(media)
-      .leftJoin(observations, eq(media.mediaID, observations.mediaID))
-      .groupBy(media.folderName)
-      .orderBy(media.folderName)
-
-    const elapsedTime = Date.now() - startTime
-    log.info(`Retrieved files data: ${rows.length} directories found in ${elapsedTime}ms`)
-    return rows
-  } catch (error) {
-    log.error(`Error querying files data: ${error.message}`)
-    throw error
-  }
-}
-
-/**
- * Get all bounding boxes for a specific media file with model provenance
- * @param {string} dbPath - Path to the SQLite database
- * @param {string} mediaID - The media ID to get bboxes for
- * @returns {Promise<Array>} - Array of observations with bbox data and model info
- */
-export async function getMediaBboxes(dbPath, mediaID, includeWithoutBbox = false) {
-  const startTime = Date.now()
-  log.info(`Querying bboxes for media: ${mediaID} (includeWithoutBbox: ${includeWithoutBbox})`)
-
-  try {
-    const studyId = getStudyIdFromPath(dbPath)
-
-    const db = await getDrizzleDb(studyId, dbPath)
-
-    // Build where clause - optionally include observations without bbox (for videos)
-    const whereClause = includeWithoutBbox
-      ? eq(observations.mediaID, mediaID)
-      : and(eq(observations.mediaID, mediaID), isNotNull(observations.bboxX))
-
-    const rows = await db
-      .select({
-        observationID: observations.observationID,
-        scientificName: observations.scientificName,
-        classificationProbability: observations.classificationProbability,
-        detectionConfidence: observations.detectionConfidence,
-        bboxX: observations.bboxX,
-        bboxY: observations.bboxY,
-        bboxWidth: observations.bboxWidth,
-        bboxHeight: observations.bboxHeight,
-        classificationMethod: observations.classificationMethod,
-        classifiedBy: observations.classifiedBy,
-        classificationTimestamp: observations.classificationTimestamp,
-        modelID: modelRuns.modelID,
-        modelVersion: modelRuns.modelVersion
-      })
-      .from(observations)
-      .leftJoin(modelOutputs, eq(observations.modelOutputID, modelOutputs.id))
-      .leftJoin(modelRuns, eq(modelOutputs.runID, modelRuns.id))
-      .where(whereClause)
-      .orderBy(desc(observations.detectionConfidence))
-
-    const elapsedTime = Date.now() - startTime
-    log.info(`Retrieved ${rows.length} bboxes for media ${mediaID} in ${elapsedTime}ms`)
-    return rows
-  } catch (error) {
-    log.error(`Error querying media bboxes: ${error.message}`)
-    throw error
-  }
-}
-
-/**
- * Get bboxes for multiple media items in a single query
- * @param {string} dbPath - Path to the SQLite database
- * @param {string[]} mediaIDs - Array of media IDs to fetch bboxes for
- * @returns {Promise<Object>} - Map of mediaID -> bboxes[]
- */
-export async function getMediaBboxesBatch(dbPath, mediaIDs) {
-  if (!mediaIDs || mediaIDs.length === 0) return {}
-
-  const startTime = Date.now()
-  log.info(`Querying bboxes for ${mediaIDs.length} media items`)
-
-  try {
-    const studyId = getStudyIdFromPath(dbPath)
-
-    const db = await getDrizzleDb(studyId, dbPath)
-
-    const rows = await db
-      .select({
-        mediaID: observations.mediaID,
-        observationID: observations.observationID,
-        scientificName: observations.scientificName,
-        classificationProbability: observations.classificationProbability,
-        detectionConfidence: observations.detectionConfidence,
-        bboxX: observations.bboxX,
-        bboxY: observations.bboxY,
-        bboxWidth: observations.bboxWidth,
-        bboxHeight: observations.bboxHeight,
-        classificationMethod: observations.classificationMethod,
-        classifiedBy: observations.classifiedBy,
-        classificationTimestamp: observations.classificationTimestamp
-      })
-      .from(observations)
-      .where(and(inArray(observations.mediaID, mediaIDs), isNotNull(observations.bboxX)))
-      .orderBy(observations.mediaID, desc(observations.detectionConfidence))
-
-    // Group results by mediaID
-    const bboxesByMedia = {}
-    for (const row of rows) {
-      if (!bboxesByMedia[row.mediaID]) {
-        bboxesByMedia[row.mediaID] = []
-      }
-      bboxesByMedia[row.mediaID].push(row)
-    }
-
-    const elapsedTime = Date.now() - startTime
-    log.info(
-      `Retrieved bboxes for ${Object.keys(bboxesByMedia).length} media items in ${elapsedTime}ms`
-    )
-    return bboxesByMedia
-  } catch (error) {
-    log.error(`Error querying media bboxes batch: ${error.message}`)
-    throw error
-  }
-}
-
-/**
- * Check if any observations with bboxes exist for the given media IDs
- * Lightweight query that returns only a boolean (uses LIMIT 1 for efficiency)
- * @param {string} dbPath - Path to the SQLite database
- * @param {string[]} mediaIDs - Array of media IDs to check
- * @returns {Promise<boolean>} - True if at least one media has bboxes
- */
-export async function checkMediaHaveBboxes(dbPath, mediaIDs) {
-  if (!mediaIDs || mediaIDs.length === 0) return false
-
-  const startTime = Date.now()
-  log.info(`Checking bbox existence for ${mediaIDs.length} media items`)
-
-  try {
-    const studyId = getStudyIdFromPath(dbPath)
-
-    const db = await getDrizzleDb(studyId, dbPath)
-
-    const result = await db
-      .select({ exists: sql`1` })
-      .from(observations)
-      .where(and(inArray(observations.mediaID, mediaIDs), isNotNull(observations.bboxX)))
-      .limit(1)
-
-    const hasBboxes = result.length > 0
-    const elapsedTime = Date.now() - startTime
-    log.info(`Bbox existence check completed in ${elapsedTime}ms: ${hasBboxes}`)
-
-    return hasBboxes
-  } catch (error) {
-    log.error(`Error checking bbox existence: ${error.message}`)
-    throw error
-  }
-}
-
-/**
- * Update media timestamp and propagate changes to related observations
- * Observations are updated with the same offset to preserve duration
- * @param {string} dbPath - Path to the SQLite database
- * @param {string} mediaID - Media ID to update
- * @param {string} newTimestamp - New timestamp in ISO 8601 format
- * @returns {Promise<Object>} - Result with success status and updated counts
- */
-export async function updateMediaTimestamp(dbPath, mediaID, newTimestamp) {
-  const startTime = Date.now()
-  log.info(`Updating timestamp for media ${mediaID} to ${newTimestamp}`)
-
-  try {
-    // Validate input parameters
-    if (!mediaID) {
-      throw new Error('Media ID is required')
-    }
-
-    if (!newTimestamp || typeof newTimestamp !== 'string') {
-      throw new Error('A valid timestamp string is required')
-    }
-
-    // Parse and validate the new timestamp
-    const newTimestampDT = DateTime.fromISO(newTimestamp)
-
-    if (!newTimestampDT.isValid) {
-      throw new Error(
-        `Invalid timestamp format: "${newTimestamp}". Please use ISO 8601 format (e.g., 2024-01-15T10:30:00.000Z)`
-      )
-    }
-
-    // Validate timestamp is within reasonable bounds (1970 to 2100)
-    const year = newTimestampDT.year
-    if (year < 1970 || year > 2100) {
-      throw new Error(`Timestamp year must be between 1970 and 2100, got ${year}`)
-    }
-
-    const studyId = getStudyIdFromPath(dbPath)
-
-    const db = await getDrizzleDb(studyId, dbPath)
-
-    // 1. Get current media timestamp
-    const currentMedia = await db
-      .select({ timestamp: media.timestamp })
-      .from(media)
-      .where(eq(media.mediaID, mediaID))
-      .get()
-
-    if (!currentMedia) {
-      throw new Error(`Media not found: ${mediaID}`)
-    }
-
-    // Handle case where current timestamp is null or invalid
-    const oldTimestamp = currentMedia.timestamp ? DateTime.fromISO(currentMedia.timestamp) : null
-
-    if (!oldTimestamp || !oldTimestamp.isValid) {
-      // If no valid old timestamp, just set the new one without offset calculation
-      log.info(`No valid existing timestamp for media ${mediaID}, setting directly`)
-
-      await db.update(media).set({ timestamp: newTimestamp }).where(eq(media.mediaID, mediaID))
-
-      // Update observations with the new timestamp directly (no offset)
-      const relatedObservations = await db
-        .select({ observationID: observations.observationID })
-        .from(observations)
-        .where(eq(observations.mediaID, mediaID))
-
-      let updatedCount = 0
-      for (const obs of relatedObservations) {
-        await db
-          .update(observations)
-          .set({ eventStart: newTimestamp })
-          .where(eq(observations.observationID, obs.observationID))
-        updatedCount++
-      }
-
-      const elapsedTime = Date.now() - startTime
-      log.info(`Set media timestamp and ${updatedCount} observations in ${elapsedTime}ms`)
-
-      return {
-        success: true,
-        mediaID,
-        newTimestamp,
-        observationsUpdated: updatedCount
-      }
-    }
-
-    // Calculate the offset in milliseconds
-    const offsetMs = newTimestampDT.toMillis() - oldTimestamp.toMillis()
-
-    // 2. Update media.timestamp - format to match original
-    const formattedNewTimestamp = formatToMatchOriginal(newTimestampDT, currentMedia.timestamp)
-    await db
-      .update(media)
-      .set({ timestamp: formattedNewTimestamp })
-      .where(eq(media.mediaID, mediaID))
-
-    // 3. Get all related observations
-    const relatedObservations = await db
-      .select({
-        observationID: observations.observationID,
-        eventStart: observations.eventStart,
-        eventEnd: observations.eventEnd
-      })
-      .from(observations)
-      .where(eq(observations.mediaID, mediaID))
-
-    // 4. Update each observation with offset-preserved times (preserving original format)
-    let updatedCount = 0
-    for (const obs of relatedObservations) {
-      const updateData = {}
-
-      // Update eventStart with offset - preserve original format
-      if (obs.eventStart) {
-        const oldEventStart = DateTime.fromISO(obs.eventStart)
-        if (oldEventStart.isValid) {
-          const newEventStart = oldEventStart.plus({ milliseconds: offsetMs })
-          updateData.eventStart = formatToMatchOriginal(newEventStart, obs.eventStart)
-        }
-      }
-
-      // Update eventEnd with SAME offset (preserving duration) - preserve original format
-      if (obs.eventEnd) {
-        const oldEventEnd = DateTime.fromISO(obs.eventEnd)
-        if (oldEventEnd.isValid) {
-          const newEventEnd = oldEventEnd.plus({ milliseconds: offsetMs })
-          updateData.eventEnd = formatToMatchOriginal(newEventEnd, obs.eventEnd)
-        }
-      }
-
-      if (Object.keys(updateData).length > 0) {
-        await db
-          .update(observations)
-          .set(updateData)
-          .where(eq(observations.observationID, obs.observationID))
-        updatedCount++
-      }
-    }
-
-    const elapsedTime = Date.now() - startTime
-    log.info(
-      `Updated media timestamp to "${formattedNewTimestamp}" and ${updatedCount} observations in ${elapsedTime}ms`
-    )
-
-    return {
-      success: true,
-      mediaID,
-      newTimestamp: formattedNewTimestamp,
-      observationsUpdated: updatedCount
-    }
-  } catch (error) {
-    log.error(`Error updating media timestamp: ${error.message}`)
-    throw error
-  }
-}
-
-/**
- * Update an observation's classification (species) with CamTrap DP compliant fields.
- * When a human updates the classification:
- * - scientificName is updated to the new value
- * - classificationMethod is set to 'human'
- * - classifiedBy is set to 'User'
- * - classificationTimestamp is set to current ISO 8601 timestamp
- * - classificationProbability is cleared (null) for human classifications per CamTrap DP spec
- *
- * @param {string} dbPath - Path to the SQLite database
- * @param {string} observationID - The observation ID to update
- * @param {Object} updates - The update values
- * @param {string} updates.scientificName - The new scientific name (can be empty for blank)
- * @param {string} [updates.commonName] - Optional common name
- * @param {string} [updates.observationType] - Optional observation type (e.g., 'blank', 'animal')
- * @returns {Promise<Object>} - The updated observation
- */
-export async function updateObservationClassification(dbPath, observationID, updates) {
-  const startTime = Date.now()
-  log.info(`Updating observation classification: ${observationID}`)
-
-  try {
-    const studyId = getStudyIdFromPath(dbPath)
-
-    const db = await getDrizzleDb(studyId, dbPath)
-
-    // Prepare update values following CamTrap DP specification
-    const updateValues = {
-      scientificName: updates.scientificName || null,
-      classificationMethod: 'human',
-      classifiedBy: 'User',
-      classificationTimestamp: new Date().toISOString(),
-      // Per CamTrap DP spec: "Omit or provide an approximate probability for human classifications"
-      // We set to null to indicate this is a human classification without probability
-      classificationProbability: null
-    }
-
-    // Add optional fields if provided
-    if (updates.commonName !== undefined) {
-      updateValues.commonName = updates.commonName
-    }
-
-    if (updates.observationType !== undefined) {
-      updateValues.observationType = updates.observationType
-    }
-
-    // Perform the update
-    await db
-      .update(observations)
-      .set(updateValues)
-      .where(eq(observations.observationID, observationID))
-
-    // Fetch and return the updated observation
-    const updatedObservation = await db
-      .select()
-      .from(observations)
-      .where(eq(observations.observationID, observationID))
-      .get()
-
-    const elapsedTime = Date.now() - startTime
-    log.info(
-      `Updated observation ${observationID} to "${updates.scientificName || 'blank'}" in ${elapsedTime}ms`
-    )
-    return updatedObservation
-  } catch (error) {
-    log.error(`Error updating observation classification: ${error.message}`)
-    throw error
-  }
-}
-
-/**
- * Update an observation's bounding box coordinates.
- * When a human updates the bbox:
- * - Bbox coordinates are updated (bboxX, bboxY, bboxWidth, bboxHeight)
- * - classificationMethod is set to 'human'
- * - classifiedBy is set to 'User'
- * - classificationTimestamp is updated
- *
- * @param {string} dbPath - Path to the SQLite database
- * @param {string} observationID - The observation ID to update
- * @param {Object} bboxUpdates - The new bbox coordinates
- * @param {number} bboxUpdates.bboxX - Left edge (0-1 normalized)
- * @param {number} bboxUpdates.bboxY - Top edge (0-1 normalized)
- * @param {number} bboxUpdates.bboxWidth - Width (0-1 normalized)
- * @param {number} bboxUpdates.bboxHeight - Height (0-1 normalized)
- * @returns {Promise<Object>} - The updated observation
- */
-export async function updateObservationBbox(dbPath, observationID, bboxUpdates) {
-  const startTime = Date.now()
-  log.info(`Updating observation bbox: ${observationID}`)
-
-  try {
-    const studyId = getStudyIdFromPath(dbPath)
-
-    const db = await getDrizzleDb(studyId, dbPath)
-
-    const { bboxX, bboxY, bboxWidth, bboxHeight } = bboxUpdates
-
-    // Validate bbox values are in valid range
-    if (
-      bboxX < 0 ||
-      bboxX > 1 ||
-      bboxY < 0 ||
-      bboxY > 1 ||
-      bboxWidth <= 0 ||
-      bboxWidth > 1 ||
-      bboxHeight <= 0 ||
-      bboxHeight > 1 ||
-      bboxX + bboxWidth > 1.001 ||
-      bboxY + bboxHeight > 1.001
-    ) {
-      throw new Error('Invalid bbox coordinates: must be normalized (0-1) and within bounds')
-    }
-
-    // Prepare update values
-    const updateValues = {
-      bboxX,
-      bboxY,
-      bboxWidth,
-      bboxHeight,
-      classificationMethod: 'human',
-      classifiedBy: 'User',
-      classificationTimestamp: new Date().toISOString()
-    }
-
-    // Perform the update
-    await db
-      .update(observations)
-      .set(updateValues)
-      .where(eq(observations.observationID, observationID))
-
-    // Fetch and return the updated observation
-    const updatedObservation = await db
-      .select()
-      .from(observations)
-      .where(eq(observations.observationID, observationID))
-      .get()
-
-    const elapsedTime = Date.now() - startTime
-    log.info(`Updated observation ${observationID} bbox in ${elapsedTime}ms`)
-    return updatedObservation
-  } catch (error) {
-    log.error(`Error updating observation bbox: ${error.message}`)
-    throw error
-  }
-}
-
-/**
- * Delete an observation from the database.
- * This permanently removes the observation record.
- *
- * @param {string} dbPath - Path to the SQLite database
- * @param {string} observationID - The observation ID to delete
- * @returns {Promise<Object>} - Success indicator with deleted observationID
- */
-export async function deleteObservation(dbPath, observationID) {
-  const startTime = Date.now()
-  log.info(`Deleting observation: ${observationID}`)
-
-  try {
-    const studyId = getStudyIdFromPath(dbPath)
-
-    const db = await getDrizzleDb(studyId, dbPath)
-
-    // Delete the observation
-    await db.delete(observations).where(eq(observations.observationID, observationID))
-
-    const elapsedTime = Date.now() - startTime
-    log.info(`Deleted observation ${observationID} in ${elapsedTime}ms`)
-    return { success: true, observationID }
-  } catch (error) {
-    log.error(`Error deleting observation: ${error.message}`)
-    throw error
-  }
-}
-
-/**
- * Create a new observation with bounding box (human-drawn).
- * Follows CamTrap DP specification for human classifications.
- *
- * @param {string} dbPath - Path to the SQLite database
- * @param {Object} observationData - The observation data
- * @param {string} observationData.mediaID - Associated media ID
- * @param {string} observationData.deploymentID - Associated deployment ID
- * @param {string} observationData.timestamp - Media timestamp (ISO 8601)
- * @param {string|null} observationData.scientificName - Species (null for unknown)
- * @param {string|null} observationData.commonName - Common name (optional)
- * @param {number} observationData.bboxX - Left edge (0-1 normalized)
- * @param {number} observationData.bboxY - Top edge (0-1 normalized)
- * @param {number} observationData.bboxWidth - Width (0-1 normalized)
- * @param {number} observationData.bboxHeight - Height (0-1 normalized)
- * @returns {Promise<Object>} - The created observation
- */
-export async function createObservation(dbPath, observationData) {
-  const startTime = Date.now()
-  log.info(`Creating new observation for media: ${observationData.mediaID}`)
-
-  try {
-    const studyId = getStudyIdFromPath(dbPath)
-
-    const db = await getDrizzleDb(studyId, dbPath)
-
-    const {
-      mediaID,
-      deploymentID,
-      timestamp,
-      scientificName,
-      commonName,
-      bboxX,
-      bboxY,
-      bboxWidth,
-      bboxHeight
-    } = observationData
-
-    // Only validate bbox if coordinates are provided (allow null for observations without bbox)
-    const hasBbox = bboxX !== null && bboxX !== undefined
-    if (hasBbox) {
-      if (
-        bboxX < 0 ||
-        bboxX > 1 ||
-        bboxY < 0 ||
-        bboxY > 1 ||
-        bboxWidth <= 0 ||
-        bboxWidth > 1 ||
-        bboxHeight <= 0 ||
-        bboxHeight > 1 ||
-        bboxX + bboxWidth > 1.001 ||
-        bboxY + bboxHeight > 1.001
-      ) {
-        throw new Error('Invalid bbox coordinates: must be normalized (0-1) and within bounds')
-      }
-    }
-
-    // Generate IDs
-    const observationID = crypto.randomUUID()
-    const eventID = crypto.randomUUID()
-
-    // Prepare observation data following CamTrap DP specification
-    const newObservation = {
-      observationID,
-      mediaID,
-      deploymentID,
-      eventID,
-      eventStart: timestamp,
-      eventEnd: timestamp,
-      scientificName: scientificName || null,
-      commonName: commonName || null,
-      observationType: 'animal',
-      classificationProbability: null, // Human classification - no classificationProbability score
-      count: 1,
-      bboxX: hasBbox ? bboxX : null,
-      bboxY: hasBbox ? bboxY : null,
-      bboxWidth: hasBbox ? bboxWidth : null,
-      bboxHeight: hasBbox ? bboxHeight : null,
-      modelOutputID: null, // No model involved
-      classificationMethod: 'human',
-      classifiedBy: 'User',
-      classificationTimestamp: new Date().toISOString()
-    }
-
-    // Insert the observation
-    await db.insert(observations).values(newObservation)
-
-    // Fetch and return the created observation
-    const createdObservation = await db
-      .select()
-      .from(observations)
-      .where(eq(observations.observationID, observationID))
-      .get()
-
-    const elapsedTime = Date.now() - startTime
-    log.info(
-      `Created observation ${observationID} for species "${scientificName || 'unknown'}" in ${elapsedTime}ms`
-    )
-    return createdObservation
-  } catch (error) {
-    log.error(`Error creating observation: ${error.message}`)
-    throw error
-  }
-}
-
-/**
- * Get all distinct species names from the observations table
- * Used to populate dropdowns for species selection
- * @param {string} dbPath - Path to the SQLite database
- * @returns {Promise<Array>} - Array of distinct species with scientificName and commonName
- */
-export async function getDistinctSpecies(dbPath) {
-  const startTime = Date.now()
-  log.info(`Querying distinct species from: ${dbPath}`)
-
-  try {
-    const studyId = getStudyIdFromPath(dbPath)
-
-    const db = await getDrizzleDb(studyId, dbPath)
-
-    const rows = await db
-      .select({
-        scientificName: observations.scientificName,
-        commonName: observations.commonName,
-        observationCount: count(observations.observationID).as('observationCount')
-      })
-      .from(observations)
-      .where(and(isNotNull(observations.scientificName), ne(observations.scientificName, '')))
-      .groupBy(observations.scientificName)
-      .orderBy(desc(count(observations.observationID)), observations.scientificName)
-
-    const elapsedTime = Date.now() - startTime
-    log.info(`Retrieved ${rows.length} distinct species in ${elapsedTime}ms`)
-    return rows
-  } catch (error) {
-    log.error(`Error querying distinct species: ${error.message}`)
-    throw error
-  }
-}
-
-/**
- * Check if a study has observations with non-null eventIDs (imported from CamtrapDP)
- * Used to determine default sequence grouping behavior in the UI
- */
-export async function checkStudyHasEventIDs(dbPath) {
-  const startTime = Date.now()
-  log.info(`Checking if study has eventIDs: ${dbPath}`)
-
-  try {
-    const studyId = getStudyIdFromPath(dbPath)
-    const db = await getDrizzleDb(studyId, dbPath)
-
-    const result = await db
-      .select({ eventID: observations.eventID })
-      .from(observations)
-      .where(and(isNotNull(observations.eventID), ne(observations.eventID, '')))
-      .limit(1)
-
-    const hasEventIDs = result.length > 0
-    const elapsedTime = Date.now() - startTime
-    log.info(`Study has eventIDs: ${hasEventIDs} (checked in ${elapsedTime}ms)`)
-    return hasEventIDs
-  } catch (error) {
-    log.error(`Error checking study eventIDs: ${error.message}`)
-    throw error
-  }
-}
-
-/**
- * Assigns sequence IDs to media candidates based on timestamp proximity within the same deployment.
- * Reuses the same logic as the media tab's groupMediaIntoSequences function.
- * Media within gapThresholdSeconds of each other at the same deployment get the same sequence ID.
- *
- * @param {Array} candidates - Array of media candidates with timestamp and deploymentID
- * @param {number} gapThresholdSeconds - Maximum gap in seconds to consider media as same sequence (default: 120)
- * @returns {Array} - Same candidates array with sequenceID added to each item
- */
-function assignSequenceIDs(candidates, gapThresholdSeconds = 120) {
-  if (!candidates || candidates.length === 0) return candidates
-
-  // Sort by deploymentID then timestamp for grouping
-  const sorted = [...candidates].sort((a, b) => {
-    const deployA = a.deploymentID || ''
-    const deployB = b.deploymentID || ''
-    if (deployA !== deployB) return deployA.localeCompare(deployB)
-    const timeA = new Date(a.timestamp).getTime() || 0
-    const timeB = new Date(b.timestamp).getTime() || 0
-    return timeA - timeB
-  })
-
-  const gapMs = gapThresholdSeconds * 1000
-  let currentSequenceId = null
-  let currentDeployment = null
-  let lastTimestamp = null
-  let sequenceCounter = 0
-
-  // Create a map of mediaID to sequenceID
-  const sequenceMap = new Map()
-
-  for (const candidate of sorted) {
-    const deployment = candidate.deploymentID || null
-    let mediaTime = null
-
-    try {
-      mediaTime = new Date(candidate.timestamp).getTime()
-      if (isNaN(mediaTime)) mediaTime = null
-    } catch {
-      mediaTime = null
-    }
-
-    // Check if we should start a new sequence
-    const differentDeployment = deployment !== currentDeployment
-    const noValidTimestamp = mediaTime === null || lastTimestamp === null
-    const gapTooLarge = mediaTime && lastTimestamp && Math.abs(mediaTime - lastTimestamp) > gapMs
-
-    if (differentDeployment || noValidTimestamp || gapTooLarge) {
-      // Start new sequence
-      sequenceCounter++
-      currentSequenceId = `seq-${sequenceCounter}`
-      currentDeployment = deployment
-    }
-
-    sequenceMap.set(candidate.mediaID, currentSequenceId)
-    if (mediaTime) lastTimestamp = mediaTime
-  }
-
-  // Add sequenceID to each candidate
-  return candidates.map((c) => ({
-    ...c,
-    sequenceID: sequenceMap.get(c.mediaID) || `seq-unique-${c.mediaID}`
-  }))
-}
-
-/**
- * Get temporal bucket for a timestamp (for diversity grouping)
- * Groups timestamps into weekly buckets to ensure temporal spread
- *
- * @param {string|null} timestamp - ISO timestamp string
- * @param {number} bucketDays - Number of days per bucket (default: 7)
- * @returns {string} - Bucket identifier (e.g., "2024-W23" for week 23 of 2024)
- */
-export function getTemporalBucket(timestamp, bucketDays = 7) {
-  if (!timestamp) return 'unknown'
-  const date = new Date(timestamp)
-  if (isNaN(date.getTime())) return 'unknown'
-  const startOfYear = new Date(date.getFullYear(), 0, 1)
-  const dayOfYear = Math.floor((date - startOfYear) / (24 * 60 * 60 * 1000))
-  const bucketIndex = Math.floor(dayOfYear / bucketDays)
-  return `${date.getFullYear()}-W${bucketIndex}`
-}
-
-/**
- * Select diverse media from candidates using a greedy algorithm
- * Applies diversity constraints to ensure variety in species, deployments, time periods, and sequences
- *
- * @param {Array} candidates - Scored media candidates (sorted by score descending), with sequenceID assigned
- * @param {number} limit - Target number of results
- * @param {Object} config - Diversity configuration
- * @param {number} config.maxPerSpecies - Max images per species (default: 2)
- * @param {number} config.maxPerDeployment - Max images per deployment (default: 3)
- * @param {number} config.maxPerTemporalBucket - Max images per time bucket (default: 4)
- * @param {number} config.maxPerSequence - Max images per sequence (default: 1)
- * @param {number} config.temporalBucketDays - Days per temporal bucket (default: 7)
- * @param {number} config.minQualityThreshold - Minimum quality score to consider (default: 0.3)
- * @returns {Array} - Selected diverse media
- */
-export function selectDiverseMedia(candidates, limit, config = {}) {
-  const {
-    maxPerSpecies = 2,
-    maxPerDeployment = 3,
-    maxPerTemporalBucket = 4,
-    maxPerSequence = 1,
-    temporalBucketDays = 7,
-    minQualityThreshold = 0.3
-  } = config
-
-  // Handle null/undefined/empty input
-  if (!candidates || candidates.length === 0) {
-    return []
-  }
-
-  // Filter by quality threshold first
-  const qualified = candidates.filter((c) => c.compositeScore >= minQualityThreshold)
-
-  const selected = []
-  const selectedIds = new Set()
-
-  // Track selection counts per dimension
-  const speciesCount = new Map()
-  const deploymentCount = new Map()
-  const temporalBucketCount = new Map()
-  const sequenceCount = new Map()
-
-  // ============================================================
-  // PHASE 1: Guarantee one image per species (highest-scoring)
-  // This ensures every species gets representation before global
-  // constraints (deployment, temporal) can block them
-  // ============================================================
-
-  // Group candidates by species
-  const speciesGroups = new Map()
-  for (const c of qualified) {
-    const species = c.scientificName || 'unknown'
-    if (!speciesGroups.has(species)) {
-      speciesGroups.set(species, [])
-    }
-    speciesGroups.get(species).push(c)
-  }
-
-  // Pick best from each species (only sequence constraint in phase 1)
-  for (const [species, items] of speciesGroups) {
-    if (selected.length >= limit) break
-
-    // Items are already sorted by score (from SQL ORDER BY)
-    for (const candidate of items) {
-      const sequenceId = candidate.sequenceID || null
-
-      // Only check sequence constraint in phase 1 - no deployment/temporal limits
-      if (sequenceId && (sequenceCount.get(sequenceId) || 0) >= maxPerSequence) continue
-
-      // Select this candidate
-      selected.push(candidate)
-      selectedIds.add(candidate.mediaID)
-      speciesCount.set(species, 1)
-
-      // Update all tracking maps for phase 2
-      const deployment = candidate.deploymentID || 'unknown'
-      const bucket = getTemporalBucket(candidate.timestamp, temporalBucketDays)
-      deploymentCount.set(deployment, (deploymentCount.get(deployment) || 0) + 1)
-      temporalBucketCount.set(bucket, (temporalBucketCount.get(bucket) || 0) + 1)
-      if (sequenceId) {
-        sequenceCount.set(sequenceId, 1)
-      }
-
-      break // Only one per species in phase 1
-    }
-  }
-
-  // ============================================================
-  // PHASE 2: Fill remaining slots with full diversity constraints
-  // Now apply all constraints including deployment and temporal
-  // ============================================================
-
-  for (const candidate of qualified) {
-    if (selected.length >= limit) break
-    if (selectedIds.has(candidate.mediaID)) continue
-
-    const species = candidate.scientificName || 'unknown'
-    const deployment = candidate.deploymentID || 'unknown'
-    const bucket = getTemporalBucket(candidate.timestamp, temporalBucketDays)
-    const sequenceId = candidate.sequenceID || null
-
-    // Check ALL diversity constraints
-    if ((speciesCount.get(species) || 0) >= maxPerSpecies) continue
-    if ((deploymentCount.get(deployment) || 0) >= maxPerDeployment) continue
-    if ((temporalBucketCount.get(bucket) || 0) >= maxPerTemporalBucket) continue
-    if (sequenceId && (sequenceCount.get(sequenceId) || 0) >= maxPerSequence) continue
-
-    // Select this candidate
-    selected.push(candidate)
-    selectedIds.add(candidate.mediaID)
-    speciesCount.set(species, (speciesCount.get(species) || 0) + 1)
-    deploymentCount.set(deployment, (deploymentCount.get(deployment) || 0) + 1)
-    temporalBucketCount.set(bucket, (temporalBucketCount.get(bucket) || 0) + 1)
-    if (sequenceId) {
-      sequenceCount.set(sequenceId, (sequenceCount.get(sequenceId) || 0) + 1)
-    }
-  }
-
-  // ============================================================
-  // PHASE 3: Fallback - relax species/deployment/temporal constraints
-  // but KEEP sequence constraint (to avoid duplicate burst images)
-  // ============================================================
-
-  if (selected.length < limit) {
-    for (const candidate of qualified) {
-      if (selected.length >= limit) break
-      if (selectedIds.has(candidate.mediaID)) continue
-
-      // Still respect sequence constraint in fallback - no duplicate bursts
-      const sequenceId = candidate.sequenceID || null
-      if (sequenceId && (sequenceCount.get(sequenceId) || 0) >= maxPerSequence) continue
-
-      selected.push(candidate)
-      selectedIds.add(candidate.mediaID)
-      if (sequenceId) {
-        sequenceCount.set(sequenceId, (sequenceCount.get(sequenceId) || 0) + 1)
-      }
-    }
-  }
-
-  return selected
-}
-
-/**
- * Get "best" media files using a hybrid approach:
- * 1. User-marked favorites first (sorted by timestamp descending)
- * 2. Auto-scored captures to fill remaining slots with diversity constraints
- *
- * Scoring formula for non-favorites (weights):
- * - 15%: Bbox area (sweet spot 10-60% of image)
- * - 20%: Fully visible (not cut off at edges)
- * - 15%: Padding (distance to nearest edge)
- * - 15%: Detection confidence
- * - 10%: Classification probability
- * - 15%: Rarity boost (rare species score higher, common species penalized)
- * - 10%: Daytime boost (favor daylight captures)
- *
- * Diversity constraints applied via post-processing:
- * - Max 2 images per species
- * - Max 3 images per deployment
- * - Max 4 images per weekly time bucket
- * - Max 1 image per event/sequence (if eventID exists)
- *
- * @param {string} dbPath - Path to the SQLite database
- * @param {Object} options - Query options
- * @param {number} options.limit - Maximum number of media to return (default: 12)
- * @returns {Promise<Array>} - Media files with favorites first, then diverse scored captures
- */
-export async function getBestMedia(dbPath, options = {}) {
-  const { limit = 12 } = options
-  const startTime = Date.now()
-  log.info(`Querying best media (hybrid mode) from: ${dbPath}`)
-
-  try {
-    const studyId = getStudyIdFromPath(dbPath)
-
-    // Step 1: Get user-marked favorites first
-    // Note: We need to join observations via both mediaID AND timestamp (for CamTrap DP datasets
-    // where observations.mediaID is NULL and link is via eventStart = media.timestamp)
-    const favoritesQuery = `
-      SELECT
-        m.mediaID,
-        m.filePath,
-        m.fileName,
-        m.timestamp,
-        m.deploymentID,
-        m.fileMediatype,
-        m.favorite,
-        COALESCE(o1.observationID, o2.observationID) as observationID,
-        COALESCE(o1.scientificName, o2.scientificName) as scientificName,
-        COALESCE(o1.bboxX, o2.bboxX) as bboxX,
-        COALESCE(o1.bboxY, o2.bboxY) as bboxY,
-        COALESCE(o1.bboxWidth, o2.bboxWidth) as bboxWidth,
-        COALESCE(o1.bboxHeight, o2.bboxHeight) as bboxHeight,
-        COALESCE(o1.detectionConfidence, o2.detectionConfidence) as detectionConfidence,
-        COALESCE(o1.classificationProbability, o2.classificationProbability) as classificationProbability,
-        999.0 as compositeScore
-      FROM media m
-      -- Strategy 1: Join via mediaID (for ML runs, Wildlife Insights, Deepfaune)
-      LEFT JOIN (
-        SELECT
-          mediaID,
-          observationID,
-          scientificName,
-          bboxX, bboxY, bboxWidth, bboxHeight,
-          detectionConfidence,
-          classificationProbability,
-          ROW_NUMBER() OVER (PARTITION BY mediaID ORDER BY detectionConfidence DESC) as rn
-        FROM observations
-        WHERE scientificName IS NOT NULL AND scientificName != ''
-          AND mediaID IS NOT NULL
-      ) o1 ON m.mediaID = o1.mediaID AND o1.rn = 1
-      -- Strategy 2: Join via timestamp (for CamTrap DP datasets where mediaID is NULL)
-      LEFT JOIN (
-        SELECT
-          eventStart,
-          observationID,
-          scientificName,
-          bboxX, bboxY, bboxWidth, bboxHeight,
-          detectionConfidence,
-          classificationProbability,
-          ROW_NUMBER() OVER (PARTITION BY eventStart ORDER BY detectionConfidence DESC) as rn
-        FROM observations
-        WHERE scientificName IS NOT NULL AND scientificName != ''
-          AND mediaID IS NULL
-      ) o2 ON m.timestamp = o2.eventStart AND o2.rn = 1
-      WHERE m.favorite = 1
-        AND COALESCE(o1.scientificName, o2.scientificName) IS NOT NULL
-      ORDER BY m.timestamp DESC
-      LIMIT ?
-    `
-
-    const favorites = await executeRawQuery(studyId, dbPath, favoritesQuery, [limit])
-    log.info(`Found ${favorites.length} favorites`)
-
-    // If we have enough favorites, return them
-    if (favorites.length >= limit) {
-      const elapsedTime = Date.now() - startTime
-      log.info(`Retrieved ${favorites.length} best media (all favorites) in ${elapsedTime}ms`)
-      return favorites
-    }
-
-    // Step 2: Get auto-scored non-favorites to fill remaining slots
-    const remainingSlots = limit - favorites.length
-    const favoriteMediaIDs = favorites.map((f) => f.mediaID)
-
-    // Use stratified sampling: fetch top N candidates per species
-    // This ensures every species has candidates for diversity selection,
-    // not just the rare species that score higher with rarity boost
-    const candidatesPerSpecies = 15
-
-    // Build exclusion clause for already-fetched favorites
-    const exclusionClause =
-      favoriteMediaIDs.length > 0
-        ? `AND m.mediaID NOT IN (${favoriteMediaIDs.map(() => '?').join(', ')})`
-        : ''
-
-    // Use raw SQL for complex scoring calculation with species rarity
-    const query = `
-      WITH
-      -- Calculate species counts for rarity scoring
-      species_counts AS (
-        SELECT scientificName, COUNT(*) as species_total
-        FROM observations
-        WHERE scientificName IS NOT NULL AND scientificName != ''
-          AND (observationType IS NULL OR observationType != 'blank')
-        GROUP BY scientificName
-      ),
-      -- Get max species count for normalization
-      max_species_count AS (
-        SELECT MAX(species_total) as max_count FROM species_counts
-      ),
-      scored_observations AS (
-        SELECT
-          o.mediaID,
-          o.observationID,
-          o.scientificName,
-          o.bboxX,
-          o.bboxY,
-          o.bboxWidth,
-          o.bboxHeight,
-          o.detectionConfidence,
-          o.classificationProbability,
-          o.eventID,
-          m.deploymentID,
-          m.timestamp,
-          -- Calculate bbox area
-          (o.bboxWidth * o.bboxHeight) as bboxArea,
-          -- Check if fully visible (1 = yes, 0 = no)
-          CASE WHEN o.bboxX >= 0 AND o.bboxY >= 0
-               AND (o.bboxX + o.bboxWidth) <= 1.0
-               AND (o.bboxY + o.bboxHeight) <= 1.0
-          THEN 1.0 ELSE 0.0 END as isFullyVisible,
-          -- Calculate padding (minimum distance to any edge)
-          MIN(o.bboxX, o.bboxY, 1.0 - o.bboxX - o.bboxWidth, 1.0 - o.bboxY - o.bboxHeight) as padding,
-          -- Calculate rarity score: rare species get higher scores, common species get lower scores
-          -- Formula: 1 - log(species_count) / log(max_count), normalized to 0-1
-          -- Common species (count ~= max) -> score ~= 0
-          -- Rare species (count << max) -> score ~= 1
-          COALESCE(
-            CASE
-              WHEN sc.species_total IS NULL THEN 0.5
-              WHEN sc.species_total <= 1 THEN 1.0
-              ELSE MAX(0.0, 1.0 - (LOG(sc.species_total + 1.0) / LOG(COALESCE((SELECT max_count FROM max_species_count), 100.0) + 1.0)))
-            END,
-            0.5
-          ) as rarityScore,
-          -- Calculate daytime score: favor captures during daylight hours (6am-6pm)
-          -- Extract hour from timestamp and score based on typical daylight
-          CASE
-            WHEN m.timestamp IS NULL THEN 0.5
-            WHEN CAST(strftime('%H', m.timestamp) AS INTEGER) BETWEEN 8 AND 16 THEN 1.0  -- Peak daylight (8am-4pm)
-            WHEN CAST(strftime('%H', m.timestamp) AS INTEGER) BETWEEN 6 AND 18 THEN 0.7  -- Extended daylight (6am-6pm)
-            ELSE 0.2  -- Night time
-          END as daytimeScore
-        FROM observations o
-        INNER JOIN media m ON o.mediaID = m.mediaID
-        LEFT JOIN species_counts sc ON o.scientificName = sc.scientificName
-        WHERE o.bboxX IS NOT NULL
-          AND o.bboxWidth IS NOT NULL
-          AND o.bboxWidth > 0
-          AND o.bboxHeight > 0
-          AND o.scientificName IS NOT NULL
-          AND o.scientificName != ''
-          -- Exclude favorites (they're already included)
-          AND (m.favorite IS NULL OR m.favorite = 0)
-          -- Exclude videos (images only)
-          AND (m.fileMediatype IS NULL OR m.fileMediatype NOT LIKE 'video/%')
-          -- Exclude blanks
-          AND (o.observationType IS NULL OR o.observationType != 'blank')
-          -- Exclude humans/persons (case-insensitive)
-          AND LOWER(o.scientificName) NOT IN ('homo sapiens', 'human', 'person', 'people')
-          AND LOWER(o.scientificName) NOT LIKE '%human%'
-          AND LOWER(o.scientificName) NOT LIKE '%person%'
-          -- Exclude vehicles
-          AND LOWER(o.scientificName) NOT IN ('vehicle', 'car', 'truck', 'motorcycle', 'bike', 'bicycle')
-          AND LOWER(o.scientificName) NOT LIKE '%vehicle%'
-          -- Exclude "other" class
-          AND LOWER(o.scientificName) != 'other'
-          ${exclusionClause}
-      ),
-      scored_with_formula AS (
-        SELECT
-          *,
-          -- Final composite score with rarity and daytime boost
-          (
-            -- Area component (15%) - sweet spot 10-60%
-            CASE
-              WHEN bboxArea < 0.05 THEN bboxArea / 0.05 * 0.3
-              WHEN bboxArea < 0.10 THEN 0.3 + (bboxArea - 0.05) / 0.05 * 0.3
-              WHEN bboxArea <= 0.60 THEN 0.6 + (bboxArea - 0.10) / 0.50 * 0.4
-              WHEN bboxArea <= 0.90 THEN 1.0 - (bboxArea - 0.60) / 0.30 * 0.3
-              ELSE 0.7 - (bboxArea - 0.90) / 0.10 * 0.4
-            END * 0.15
-            -- Visibility component (20%)
-            + isFullyVisible * 0.20
-            -- Padding component (15%), capped at padding >= 0.20
-            + MIN(MAX(padding, 0) * 5, 1.0) * 0.15
-            -- Detection confidence (15%)
-            + COALESCE(detectionConfidence, 0.5) * 0.15
-            -- Classification probability (10%)
-            + COALESCE(classificationProbability, 0.5) * 0.10
-            -- Rarity boost (15%) - rare species get higher scores
-            + rarityScore * 0.15
-            -- Daytime boost (10%) - favor daylight captures
-            + daytimeScore * 0.10
-          ) as compositeScore
-        FROM scored_observations
-      ),
-      -- Get best observation per media (avoid duplicates)
-      best_per_media AS (
-        SELECT
-          mediaID,
-          observationID,
-          scientificName,
-          bboxX, bboxY, bboxWidth, bboxHeight,
-          detectionConfidence, classificationProbability,
-          eventID,
-          deploymentID, timestamp,
-          compositeScore,
-          ROW_NUMBER() OVER (PARTITION BY mediaID ORDER BY compositeScore DESC) as rn
-        FROM scored_with_formula
-      ),
-      -- Filter to one observation per media
-      unique_media AS (
-        SELECT * FROM best_per_media WHERE rn = 1
-      ),
-      -- Stratified sampling: rank within each species to ensure all species are represented
-      -- This prevents rare species from dominating due to rarity boost
-      ranked_per_species AS (
-        SELECT
-          *,
-          ROW_NUMBER() OVER (PARTITION BY scientificName ORDER BY compositeScore DESC) as species_rank
-        FROM unique_media
-      )
-      SELECT
-        m.mediaID,
-        m.filePath,
-        m.fileName,
-        m.timestamp,
-        m.deploymentID,
-        m.fileMediatype,
-        m.favorite,
-        r.observationID,
-        r.scientificName,
-        r.bboxX,
-        r.bboxY,
-        r.bboxWidth,
-        r.bboxHeight,
-        r.detectionConfidence,
-        r.classificationProbability,
-        r.eventID,
-        r.compositeScore
-      FROM ranked_per_species r
-      INNER JOIN media m ON r.mediaID = m.mediaID
-      WHERE r.species_rank <= ?
-      ORDER BY r.compositeScore DESC
-    `
-
-    // Build query parameters: favoriteMediaIDs (for exclusion) + candidatesPerSpecies (for stratified sampling)
-    const queryParams = [...favoriteMediaIDs, candidatesPerSpecies]
-    const scoredCandidates = await executeRawQuery(studyId, dbPath, query, queryParams)
-
-    // Assign sequence IDs using 120s gap threshold (same as media tab)
-    // This groups media by timestamp proximity within same deployment
-    const candidatesWithSequences = assignSequenceIDs(scoredCandidates, 120)
-
-    // Step 3: Apply diversity selection to scored candidates
-    // This ensures variety in species, deployments, time periods, and sequences
-    const diverseResults = selectDiverseMedia(candidatesWithSequences, remainingSlots, {
-      maxPerSpecies: 2,
-      maxPerDeployment: 3,
-      maxPerTemporalBucket: 4,
-      maxPerSequence: 1, // No duplicate images from same sequence/burst
-      temporalBucketDays: 7,
-      minQualityThreshold: 0.3
-    })
-
-    // Step 4: Combine favorites + diverse scored results
-    const combinedResults = [...favorites, ...diverseResults]
-
-    const elapsedTime = Date.now() - startTime
-    log.info(
-      `Retrieved ${combinedResults.length} best media (${favorites.length} favorites + ${diverseResults.length} diverse from ${scoredCandidates.length} candidates) in ${elapsedTime}ms`
-    )
-
-    return combinedResults
-  } catch (error) {
-    log.error(`Error querying best media: ${error.message}`)
-    throw error
-  }
-}
-
-/**
- * Update media favorite status
- * @param {string} dbPath - Path to the SQLite database
- * @param {string} mediaID - Media ID to update
- * @param {boolean} favorite - New favorite status
- * @returns {Promise<Object>} - Result with success status
- */
-export async function updateMediaFavorite(dbPath, mediaID, favorite) {
-  const startTime = Date.now()
-  log.info(`Updating favorite status for media ${mediaID} to ${favorite}`)
-
-  try {
-    // Validate input parameters
-    if (!mediaID) {
-      throw new Error('Media ID is required')
-    }
-
-    if (typeof favorite !== 'boolean') {
-      throw new Error('Favorite must be a boolean value')
-    }
-
-    const studyId = getStudyIdFromPath(dbPath)
-
-    const db = await getDrizzleDb(studyId, dbPath)
-
-    // Check if media exists
-    const existingMedia = await db
-      .select({ mediaID: media.mediaID })
-      .from(media)
-      .where(eq(media.mediaID, mediaID))
-      .get()
-
-    if (!existingMedia) {
-      throw new Error(`Media not found: ${mediaID}`)
-    }
-
-    // Update the favorite status
-    await db.update(media).set({ favorite }).where(eq(media.mediaID, mediaID))
-
-    const elapsedTime = Date.now() - startTime
-    log.info(`Updated favorite status for media ${mediaID} in ${elapsedTime}ms`)
-
-    return { success: true, mediaID, favorite }
-  } catch (error) {
-    log.error(`Error updating media favorite: ${error.message}`)
-    throw error
-  }
-}
-
-/**
- * Count media files with null timestamps
- * @param {string} dbPath - Path to the SQLite database
- * @returns {Promise<number>} - Count of media files with null timestamps
- */
-export async function countMediaWithNullTimestamps(dbPath) {
-  const startTime = Date.now()
-  log.info(`Counting media with null timestamps from: ${dbPath}`)
-
-  try {
-    const studyId = getStudyIdFromPath(dbPath)
-
-    const db = await getDrizzleDb(studyId, dbPath)
-
-    const result = await db
-      .select({ count: count().as('count') })
-      .from(media)
-      .where(isNull(media.timestamp))
-      .get()
-
-    const nullCount = result?.count || 0
-    const elapsedTime = Date.now() - startTime
-    log.info(`Found ${nullCount} media with null timestamps in ${elapsedTime}ms`)
-
-    return nullCount
-  } catch (error) {
-    log.error(`Error counting media with null timestamps: ${error.message}`)
-    throw error
-  }
-}
