+++ test/metadata-validation.test.js
@@ -8,7 +8,7 @@ import {
   metadataUpdateSchema,
   metadataCreateSchema,
   contributorRoles
-} from '../src/main/db/schemas.js'
+} from '../src/main/database/validators.js'
 
 describe('Metadata Zod Validation', () => {
   describe('contributorSchema', () => {
